sync

+++ src/example/expr.rs
@@ -4,7 +4,7 @@
         10 + 10
     }
 
-    fn foo(qq: i32) {
+    fn foo(qq: f32) {
         local_1 = 3.3 + qq;
         bar = 10.0;
         g = 9.0 + 4.0 * local_1 + 2.0 / 3.0 - 99.0 * bar;
